Refactored index.html, added functionality for reseting the game.
Created a base images folder for images that are not to be copied to the build folder.

KITTENS AND PUPPIES ADDED! :D:D:D:D:D:D:D:D:D:D:D:D:D:D

Plenty of functionality added to the controller, including checking for errors before contacting the server.